# openapi/wokcore-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WokCore Local Control Plane",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:{port}",
      "variables": {
        "port": {
          "default": "43127"
        }
      }
    }
  ],
  "paths": {
...
    "/wokcore/v1/capabilities": {
      "get": {
        "operationId": "getCapabilities",
        "security": [],
        "responses": {
          "200": {
            "description": "Versioned capability handshake",
            "headers": {
              "X-Request-Id": {"$ref": "#/components/headers/XRequestId"},
              "Cache-Control": {"$ref": "#/components/headers/CacheControl"},
              "X-Content-Type-Options": {"$ref": "#/components/headers/XContentTypeOptions"}
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Capabilities"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/wokcore/v1/service/status": {
      "get": {
        "operationId": "getServiceStatus",
        "security": [
          {
            "managementBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Service lifecycle status",
            "headers": {
              "X-Request-Id": {"$ref": "#/components/headers/XRequestId"},
              "Cache-Control": {"$ref": "#/components/headers/CacheControl"},
              "X-Content-Type-Options": {"$ref": "#/components/headers/XContentTypeOptions"}
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lifecycle"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/wokcore/v1/service/drain": {
      "post": {
        "operationId": "beginServiceDrain",
        "security": [
          {
            "managementBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Drain state",
            "headers": {
              "X-Request-Id": {"$ref": "#/components/headers/XRequestId"},
              "Cache-Control": {"$ref": "#/components/headers/CacheControl"},
              "X-Content-Type-Options": {"$ref": "#/components/headers/XContentTypeOptions"}
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lifecycle"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/wokcore/v1/service/drain/cancel": {
      "post": {
        "operationId": "cancelServiceDrain",
        "security": [
          {
            "managementBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Running state",
            "headers": {
              "X-Request-Id": {"$ref": "#/components/headers/XRequestId"},
              "Cache-Control": {"$ref": "#/components/headers/CacheControl"},
              "X-Content-Type-Options": {"$ref": "#/components/headers/XContentTypeOptions"}
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lifecycle"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/wokcore/v1/service/stop": {
      "post": {
        "operationId": "stopService",
        "security": [
          {
            "managementBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Stopping state",
            "headers": {
              "X-Request-Id": {"$ref": "#/components/headers/XRequestId"},
              "Cache-Control": {"$ref": "#/components/headers/CacheControl"},
              "X-Content-Type-Options": {"$ref": "#/components/headers/XContentTypeOptions"}
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lifecycle"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/wokcore/v1/clients/authorize": {
      "post": {
        "operationId": "authorizeClient",
        "security": [
          {
            "managementBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorizeRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "One-time proxy token response",
            "headers": {
              "X-Request-Id": {"$ref": "#/components/headers/XRequestId"},
              "Cache-Control": {"$ref": "#/components/headers/CacheControl"},
              "X-Content-Type-Options": {"$ref": "#/components/headers/XContentTypeOptions"}
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorizeResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/wokcore/v1/sessions": {
      "get": {
        "operationId": "listSessions",
        "security": [
          {"managementBearer": []},
          {"clientBearer": []}
        ],
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "schema": {"type": "string", "enum": ["codex", "claude", "gemini"]}
          },
          {
            "name": "availability",
            "in": "query",
            "schema": {"type": "string", "enum": ["available", "unavailable"]}
          },
          {
            "name": "before",
            "in": "query",
            "schema": {"type": "string", "minLength": 1}
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {"type": "integer", "minimum": 1, "maximum": 200, "default": 50}
          }
        ],
        "responses": {
          "200": {
            "description": "Bounded indexed Session page",
            "headers": {
              "X-Request-Id": {"$ref": "#/components/headers/XRequestId"},
              "Cache-Control": {"$ref": "#/components/headers/CacheControl"},
              "X-Content-Type-Options": {"$ref": "#/components/headers/XContentTypeOptions"}
            },
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/SessionList"}
              }
            }
          },
          "default": {"$ref": "#/components/responses/Error"}
        }
      }
    },
...
    "/wokcore/v1/usage": {
      "get": {
        "operationId": "getUsage",
        "security": [
          {"managementBearer": []},
          {"clientBearer": []}
        ],
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "schema": {"type": "string", "enum": ["codex", "claude", "gemini"]}
          },
          {
            "name": "session_key",
            "in": "query",
            "schema": {"$ref": "#/components/schemas/OpaqueKey"}
          },
          {
            "name": "since",
            "in": "query",
            "schema": {"$ref": "#/components/schemas/UtcTimestamp"}
          },
          {
            "name": "until",
            "in": "query",
            "schema": {"$ref": "#/components/schemas/UtcTimestamp"}
          },
          {
            "name": "group_by",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["day", "source", "model"],
              "default": "day"
            }
          },
          {
            "name": "after",
            "in": "query",
            "schema": {"type": "string", "minLength": 1}
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {"type": "integer", "minimum": 1, "maximum": 500, "default": 100}
          }
        ],
        "responses": {
          "200": {
            "description": "Bounded Session-derived usage aggregates",
            "headers": {
              "X-Request-Id": {"$ref": "#/components/headers/XRequestId"},
              "Cache-Control": {"$ref": "#/components/headers/CacheControl"},
              "X-Content-Type-Options": {"$ref": "#/components/headers/XContentTypeOptions"}
            },
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/UsageResponse"}
              }
            }
          },
          "default": {"$ref": "#/components/responses/Error"}
        }
      }
    },
    "/wokcore/v1/logs": {
      "get": {
        "operationId": "listDiagnosticEvents",
        "security": [
          {"managementBearer": []},
          {"clientBearer": []}
        ],
        "parameters": [
          {"name": "request_id", "in": "query", "schema": {"type": "string", "maxLength": 256}},
          {"name": "trace_id", "in": "query", "schema": {"type": "string", "maxLength": 256}},
          {"name": "session_key", "in": "query", "schema": {"type": "string", "maxLength": 256}},
          {
            "name": "level_min",
            "in": "query",
            "schema": {"type": "string", "enum": ["debug", "info", "warn", "error"]}
          },
          {"name": "component", "in": "query", "schema": {"type": "string"}},
          {"name": "since", "in": "query", "schema": {"$ref": "#/components/schemas/UtcTimestamp"}},
          {"name": "until", "in": "query", "schema": {"$ref": "#/components/schemas/UtcTimestamp"}},
          {
            "name": "order",
            "in": "query",
            "schema": {"type": "string", "enum": ["asc", "desc"], "default": "desc"}
          },
          {"name": "after", "in": "query", "schema": {"type": "string", "minLength": 1}},
          {
            "name": "limit",
            "in": "query",
            "schema": {"type": "integer", "minimum": 1, "maximum": 1000, "default": 100}
          }
        ],
        "responses": {
          "200": {
            "description": "Bounded redacted diagnostic event page",
            "headers": {
              "X-Request-Id": {"$ref": "#/components/headers/XRequestId"},
              "Cache-Control": {"$ref": "#/components/headers/CacheControl"},
              "X-Content-Type-Options": {"$ref": "#/components/headers/XContentTypeOptions"}
            },
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/DiagnosticLogs"}
              }
            }
          },
          "default": {"$ref": "#/components/responses/Error"}
        }
      }
    },
    "/wokcore/v1/diagnostics/export": {
      "get": {
        "operationId": "exportDiagnostics",
        "security": [
          {"managementBearer": []},
          {"clientBearer": []}
        ],
        "parameters": [
          {"name": "request_id", "in": "query", "schema": {"type": "string", "maxLength": 256}},
          {"name": "trace_id", "in": "query", "schema": {"type": "string", "maxLength": 256}},
          {"name": "session_key", "in": "query", "schema": {"type": "string", "maxLength": 256}},
          {"name": "since", "in": "query", "schema": {"$ref": "#/components/schemas/UtcTimestamp"}},
          {"name": "until", "in": "query", "schema": {"$ref": "#/components/schemas/UtcTimestamp"}},
          {"name": "include_snapshots", "in": "query", "schema": {"type": "boolean", "default": false}},
          {
            "name": "max_bytes",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 65536,
              "maximum": 67108864,
              "default": 16777216
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Validated streamed diagnostic support package",
            "headers": {
              "X-Request-Id": {"$ref": "#/components/headers/XRequestId"},
              "Cache-Control": {"$ref": "#/components/headers/CacheControl"},
              "X-Content-Type-Options": {"$ref": "#/components/headers/XContentTypeOptions"},
              "Content-Disposition": {"$ref": "#/components/headers/ContentDisposition"}
            },
            "content": {
              "application/zip": {
                "schema": {"type": "string", "contentEncoding": "binary"}
              }
            }
          },
          "default": {"$ref": "#/components/responses/Error"}
        }
      }
    },
...
  },
  "components": {
    "headers": {
      "XRequestId": {
        "description": "Fresh server-generated request UUID",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "CacheControl": {
        "description": "Disables response storage",
        "schema": {
          "type": "string",
          "const": "no-store"
        }
      },
      "XContentTypeOptions": {
        "description": "Disables content-type sniffing",
        "schema": {
          "type": "string",
          "const": "nosniff"
        }
      },
      "ContentDisposition": {
        "description": "Sanitized UTC diagnostic package filename",
        "schema": {
          "type": "string",
          "pattern": "^attachment; filename=\"wokcore-diagnostics-[0-9]{8}T[0-9]{6}Z\\.zip\"$"
        }
      }
    },
    "securitySchemes": {
      "managementBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "WokCore management token"
      },
      "clientBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Scoped WokCore client token"
      }
    },
    "responses": {
      "Error": {
        "description": "Stable control-plane error",
        "headers": {
          "X-Request-Id": {"$ref": "#/components/headers/XRequestId"},
          "Cache-Control": {"$ref": "#/components/headers/CacheControl"},
          "X-Content-Type-Options": {"$ref": "#/components/headers/XContentTypeOptions"}
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    },
    "schemas": {
      "ClientId": {
        "type": "string",
        "minLength": 1,
        "maxLength": 128,
        "pattern": "^[a-z0-9._-]*[a-z0-9][a-z0-9._-]*$"
      },
...
      "Capabilities": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "wokcore_version",
          "management_api_major",
          "minimum_management_api_major",
          "maximum_management_api_major",
          "provider_protocols",
          "capabilities",
          "instance_id"
        ],
        "properties": {
          "wokcore_version": {
            "type": "string"
          },
          "management_api_major": {
            "const": 1
          },
          "minimum_management_api_major": {
            "const": 1
          },
          "maximum_management_api_major": {
            "const": 1
          },
          "provider_protocols": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": true
          },
          "capabilities": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": true
          },
          "instance_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "Lifecycle": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "phase",
          "active_requests"
        ],
        "properties": {
          "phase": {
            "type": "string",
            "enum": [
              "starting",
              "running",
              "draining",
              "awaiting_cancellation",
              "stopping"
            ]
          },
          "active_requests": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "AuthorizeRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "client_id"
        ],
        "properties": {
          "client_id": {
            "$ref": "#/components/schemas/ClientId"
          },
          "scopes": {
            "$ref": "#/components/schemas/ClientTokenScopes"
          }
        }
      },
      "AuthorizeResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "client_id",
          "token_id",
          "token",
          "scopes"
        ],
        "properties": {
          "client_id": {
            "$ref": "#/components/schemas/ClientId"
          },
          "token_id": {
            "type": "string",
            "format": "uuid"
          },
          "token": {
            "type": "string",
            "readOnly": true
          },
          "scopes": {
            "$ref": "#/components/schemas/ClientTokenScopes"
          }
        }
      },
      "ClientTokenScopes": {
        "type": "array",
        "minItems": 1,
        "maxItems": 5,
        "uniqueItems": true,
        "items": {
          "type": "string",
          "enum": [
            "proxy.use",
            "sessions.read",
            "usage.read",
            "diagnostics.read",
            "diagnostics.export"
          ]
        }
      },
      "OpaqueKey": {
        "type": "string",
        "pattern": "^[a-f0-9]{64}$"
      },
      "UtcTimestamp": {
        "type": "string",
        "pattern": "^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}Z$"
      },
      "SessionList": {
        "type": "object",
        "additionalProperties": false,
        "required": ["schema_version", "items", "next_cursor", "index_status"],
        "properties": {
          "schema_version": {"const": 1},
          "items": {
            "type": "array",
            "maxItems": 200,
            "items": {"$ref": "#/components/schemas/SessionListItem"}
          },
          "next_cursor": {"type": ["string", "null"]},
          "index_status": {"$ref": "#/components/schemas/IndexStatus"}
        }
      },
      "SessionListItem": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "session_key",
          "source",
          "created_at",
          "last_active_at",
          "availability",
          "message_count",
          "usage_event_count"
        ],
        "properties": {
          "session_key": {"$ref": "#/components/schemas/OpaqueKey"},
          "source": {"type": "string", "enum": ["codex", "claude", "gemini"]},
          "created_at": {"$ref": "#/components/schemas/UtcTimestamp"},
          "last_active_at": {"$ref": "#/components/schemas/UtcTimestamp"},
          "availability": {"type": "string", "enum": ["available", "unavailable"]},
          "message_count": {"type": "integer", "minimum": 0},
          "usage_event_count": {"type": "integer", "minimum": 0},
          "title": {"type": "string"}
        }
      },
      "IndexStatus": {
        "type": "object",
        "additionalProperties": false,
        "required": ["phase", "sources"],
        "properties": {
          "phase": {"type": "string", "enum": ["starting", "scanning", "idle"]},
          "sources": {
            "type": "array",
            "maxItems": 3,
            "items": {"$ref": "#/components/schemas/SourceIndexStatus"}
          }
        }
      },
      "SourceIndexStatus": {
        "type": "object",
        "additionalProperties": false,
        "required": ["source", "status"],
        "properties": {
          "source": {"type": "string", "enum": ["codex", "claude", "gemini"]},
          "status": {
            "type": "string",
            "enum": [
              "undiscovered",
              "available",
              "stale",
              "unavailable",
              "resource_limited"
            ]
          },
          "last_transition_at": {"$ref": "#/components/schemas/UtcTimestamp"},
          "error_code": {
            "type": "string",
            "enum": [
              "source_root_missing",
              "source_root_unreadable",
              "source_sessions_absent",
              "source_entry_unsafe",
              "source_io_failed",
              "source_parse_invalid",
              "source_record_too_large",
              "source_replay_parent_missing",
              "source_replay_parent_ambiguous",
              "source_replay_inconsistent",
              "source_replay_limit",
              "source_candidate_interrupted"
            ]
          }
        }
      },
...
      "UsageResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["schema_version", "group_by", "totals", "buckets", "next_cursor"],
        "properties": {
          "schema_version": {"const": 1},
          "group_by": {"type": "string", "enum": ["day", "source", "model"]},
          "totals": {"$ref": "#/components/schemas/UsageTotals"},
          "buckets": {
            "type": "array",
            "maxItems": 500,
            "items": {"$ref": "#/components/schemas/UsageBucket"}
          },
          "next_cursor": {"type": ["string", "null"]}
        }
      },
      "UsageTotals": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "input_tokens",
          "output_tokens",
          "cache_read_tokens",
          "cache_write_tokens",
          "reasoning_tokens",
          "session_count"
        ],
        "properties": {
          "input_tokens": {"type": "integer", "minimum": 0},
          "output_tokens": {"type": "integer", "minimum": 0},
          "cache_read_tokens": {"type": "integer", "minimum": 0},
          "cache_write_tokens": {"type": "integer", "minimum": 0},
          "reasoning_tokens": {"type": "integer", "minimum": 0},
          "session_count": {"type": "integer", "minimum": 0}
        }
      },
      "UsageBucket": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "key",
          "input_tokens",
          "output_tokens",
          "cache_read_tokens",
          "cache_write_tokens",
          "reasoning_tokens",
          "session_count"
        ],
        "properties": {
          "key": {"type": "string"},
          "start": {"$ref": "#/components/schemas/UtcTimestamp"},
          "end": {"$ref": "#/components/schemas/UtcTimestamp"},
          "input_tokens": {"type": "integer", "minimum": 0},
          "output_tokens": {"type": "integer", "minimum": 0},
          "cache_read_tokens": {"type": "integer", "minimum": 0},
          "cache_write_tokens": {"type": "integer", "minimum": 0},
          "reasoning_tokens": {"type": "integer", "minimum": 0},
          "session_count": {"type": "integer", "minimum": 0}
        }
      },
      "DiagnosticLogs": {
        "type": "object",
        "additionalProperties": false,
        "required": ["schema_version", "items", "next_cursor", "dropped_events"],
        "properties": {
          "schema_version": {"const": 1},
          "items": {
            "type": "array",
            "maxItems": 1000,
            "items": {"type": "object"}
          },
          "next_cursor": {"type": ["string", "null"]},
          "dropped_events": {"type": "integer", "minimum": 0}
        }
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/Error"
          }
        }
      },
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "properties": {
          "code": {
            "type": "string",
            "pattern": "^[a-z][a-z0-9_]*$"
          },
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      }
    }
  }
}